Se realiza modularización del gestor de tareas con las siguientes características carpetas api services ui utils
