+++ actividades/frontend/index.html
@@ -53,6 +53,6 @@
     </section>
   </div>
 
-  <script src="./logica.js" type="module"></script>
+  <script src="./main.js" type="module"></script>
 </body>
 </html>
